+++ lib/src/engine/shadow_map.dart
@@ -9,39 +9,58 @@ class M3ShadowMap {
   RenderingContext get gl => M3AppEngine.instance.renderEngine.gl;
   static final _prog = M3AppEngine.instance.renderEngine.programSimple!;
 
-  late M3Framebuffer shadowBuffer;
-  Matrix4 matShadow = Matrix4.identity();
-  WebGLTexture get depthTexture => shadowBuffer.depthTexture;
-  M3Texture? _texDepth;
-
-  // depth image for debug
-
+  final M3Framebuffer _framebuffer;
+  int get mapW => _framebuffer.frameW;
+  int get mapH => _framebuffer.frameH;
+  WebGLTexture get depthTexture => _framebuffer.depthTexture;
+  M3Texture? _tex;
   M3Texture get texDepth {
-    _texDepth ??= M3Texture.fromWebGLTexture(depthTexture, texW: shadowBuffer.frameW, texH: shadowBuffer.frameH);
-    return _texDepth!;
+    return _tex ??= M3Texture.fromWebGLTexture(depthTexture, texW: mapW, texH: mapH);
   }
 
-  M3ShadowMap(int width, int height) {
+  M3ShadowMap(int width, int height) : _framebuffer = M3Framebuffer(width, height) {
     debugPrint('create M3ShadowMap: $width x $height');
-    shadowBuffer = M3Framebuffer(width, height);
   }
 
   void dispose() {
-    shadowBuffer.dispose();
+    _tex?.dispose();
+    _framebuffer.dispose();
   }
 
   void renderDepthPass(M3Scene scene, M3Light light) {
-    shadowBuffer.bind();
+    _framebuffer.bind();
     gl.clear(WebGL.DEPTH_BUFFER_BIT);
     gl.disable(WebGL.BLEND);
     gl.enable(WebGL.POLYGON_OFFSET_FILL);
-    // render back-face to avoid shadow acne
-    gl.frontFace(WebGL.CW);
-    gl.polygonOffset(.3, .2);
+    // render front-face and positive offset to avoid shadow acne
+    // gl.frontFace(WebGL.CCW);
+    // gl.polygonOffset(.3, .2);
+    gl.polygonOffset(1.1, 4.0);
 
-    scene.render(_prog, light);
+    if (light.isDirectional) {
+      light.updateShadowCascades(scene.cameras[0]);
+    }
+    // check if use cascaded shadow map
+    if (light.cascades.isNotEmpty) {
+      // cascaded shadow mapping
+      final backupMatrix = light.projectionMatrix;
+      for (final cascade in light.cascades) {
+        // viewport for the cascaded-shadow
+        final int y = (cascade.atlasBiasV * mapH).toInt();
+        final int height = (cascade.atlasScaleV * mapH).toInt();
+        gl.viewport(0, y, mapW, height);
+        light.projectionMatrix = cascade.projectionMatrix;
+        // frustum matrix for culling
+        light.updateFrustum(light.projectionMatrix * light.viewMatrix);
+        // render scene
+        scene.render(_prog, light);
+      }
+      light.projectionMatrix = backupMatrix;
+      light.updateFrustum(light.projectionMatrix * light.viewMatrix);
+    } else {
+      scene.render(_prog, light);
+    }
 
-    gl.frontFace(WebGL.CCW);
     gl.polygonOffset(0, 0);
     gl.disable(WebGL.POLYGON_OFFSET_FILL);
     gl.enable(WebGL.BLEND);