+++ lib/src/scene/entity.dart
@@ -9,6 +9,7 @@ import 'transform.dart';
 /// Combines a mesh, transform, color, and optional rigid body for physics simulation.
 class M3Entity {
   final M3Transform _transform = M3Transform();
+  M3Transform get transform => _transform;
   oimo.RigidBody? rigidBody;
   M3Mesh? mesh;
   Vector4 color = Vector4(1.0, 1.0, 1.0, 1.0); // RGBA