+++ test/transform_test.dart
@@ -2,8 +2,6 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:macbear_3d/macbear_3d.dart';
 
-import '../lib/src/scene/transform.dart';
-
 void main() {
   group('M3Transform', () {
     test('local to world matrix recomputation', () {